+++ R/data-stockAssessmentSummary.R
@@ -1,5 +1,10 @@
 #' Reference points, contact info, assessment metadata for USA assessed stocks
 #'
+#' @description
+#' `r lifecycle::badge("deprecated")`
+#'
+#' `stockAssessmentSummary` is deprecated and will be removed in the next major release (1.0.0). Camel case is being replaced by snake case. Please use [stock_assessment_summary] instead.
+#'
 #' Stock assessment metadata and reference points for species found in waters
 #'  of the USA. Data is provided for actively managed stocks of a federal
 #'   Fisheries Management Plan (FMP). Stocks solely managed by US States are
@@ -94,5 +99,5 @@
 #' 8. Download all Parts
 #'
 #'
-#'
+#' @keywords internal
 "stockAssessmentSummary"